refactor: 파일 구조
