+++ src/page/main/index/Main.tsx
@@ -1,12 +1,8 @@
-import { mainContainer, tttt } from "./main.css";
 /*
 localhost:5173/main
 */
 const Main = () => {
-  return (
-    <div className={mainContainer}>
-      <h1 className={tttt}>Main Page</h1>
-    </div>
-  );
+  return <div>Main</div>;
 };
+
 export default Main;
