+++ src/style/index.ts
@@ -1,3 +1,3 @@
 export * from "./reset.css";
 export * from "./global.css";
-export * from "./token/token";
+export * from "./styles.css";
