--- src/style/token/token.ts
@@ -1,3 +1,0 @@
-export * from "./color.css";
-export * from "./semanticColor.css";
-export * from "./typeStyle.css";
